+++ projects/packages/forms/src/blocks/field-consent/index.js
@@ -1,7 +1,7 @@
 import { Path } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { getIconColor } from '../contact-form/util/block-icons';
-import renderMaterialIcon from '../contact-form/util/render-material-icon';
+import renderMaterialIcon from '../shared/components/render-material-icon';
 import defaultSettings from '../shared/settings';
 import deprecated from './deprecated';
 import edit from './edit';
@@ -10,7 +10,7 @@ import save from './save';
 const name = 'field-consent';
 const settings = {
 	...defaultSettings,
-	title: __( 'Terms Consent', 'jetpack-forms' ),
+	title: __( 'Terms consent', 'jetpack-forms' ),
 	keywords: [ __( 'Consent', 'jetpack-forms' ) ],
 	description: __(
 		'Communicate site terms and offer visitors consent to those terms.',