+++ projects/packages/masterbar/src/class-main.php
@@ -14,7 +14,7 @@ use Automattic\Jetpack\Status\Host;
  */
 class Main {
 
-	const PACKAGE_VERSION = '0.17.7';
+	const PACKAGE_VERSION = '0.17.9';
 
 	/**
 	 * Initializer.