+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.13.2] - 2025-06-09
+### Fixed
+- Tests: Ensure method param count matches PHPUnit data providers. [#43828]
+
+## [6.13.1] - 2025-06-06
+### Fixed
+- Tests: Resolve warnings in PHPUnit 12.2. [#43815]
+
 ## [6.13.0] - 2025-06-05
 ### Changed
 - Limit access to manage connection dialog on WoA sites. [#43777]
@@ -1456,6 +1464,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[6.13.2]: https://github.com/Automattic/jetpack-connection/compare/v6.13.1...v6.13.2
+[6.13.1]: https://github.com/Automattic/jetpack-connection/compare/v6.13.0...v6.13.1
 [6.13.0]: https://github.com/Automattic/jetpack-connection/compare/v6.12.0...v6.13.0
 [6.12.0]: https://github.com/Automattic/jetpack-connection/compare/v6.11.9...v6.12.0
 [6.11.9]: https://github.com/Automattic/jetpack-connection/compare/v6.11.8...v6.11.9