+++ projects/packages/masterbar/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.9] - 2025-06-09
+### Changed
+- Forms: Update Feedback menu slug to point to forms inbox. [#43780]
+
+## [0.17.8] - 2025-06-06
+### Fixed
+- Fix the Jetpack submenu order for WordPress.com sites that use Nav Unification. [#43781]
+
 ## [0.17.7] - 2025-06-05
 ### Changed
 - Update dependencies. [#43785]
@@ -347,6 +355,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Notifications: Change Icon [#37676]
 - Updated package dependencies. [#37669] [#37706]
 
+[0.17.9]: https://github.com/Automattic/jetpack-masterbar/compare/v0.17.8...v0.17.9
+[0.17.8]: https://github.com/Automattic/jetpack-masterbar/compare/v0.17.7...v0.17.8
 [0.17.7]: https://github.com/Automattic/jetpack-masterbar/compare/v0.17.6...v0.17.7
 [0.17.6]: https://github.com/Automattic/jetpack-masterbar/compare/v0.17.5...v0.17.6
 [0.17.5]: https://github.com/Automattic/jetpack-masterbar/compare/v0.17.4...v0.17.5